update osgMain for good ration in screen
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@5050 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: b90ccfd0d2daf6572b50e1fa8818ccbd57f187a4
Former-commit-id: 66b0cddbe1dfd00fe657124ebb5863c88818c467

+++ src/modules/graphic/osggraph/OsgMain.cpp
@@ -465,7 +465,7 @@ initView(int x, int y, int width, int height, int /* flag */, void *screen)
     m_sceneViewer->getCamera()->setName("Cam one");
     m_sceneViewer->getCamera()->setViewport(new osg::Viewport(0, 0, grWinw, grWinh));
     m_sceneViewer->getCamera()->setGraphicsContext(gw.get());
-    m_sceneViewer->getCamera()->setProjectionMatrixAsPerspective(67.5f, static_cast<double>(grWinw / grWinh), 0.2f, 9000.0f);
+    m_sceneViewer->getCamera()->setProjectionMatrixAsPerspective(67.5f, static_cast<double>((float)grWinw / (float)grWinh), 0.2f, 9000.0f);
     //m_sceneViewer->setThreadingModel(osgViewer::Viewer::CullThreadPerCameraDrawThreadPerContext);
     m_sceneViewer->realize();
